Fix unit tests after merge of develop branch
