+++ front-end/src/app/shared/services/report-f3x.service.ts
@@ -64,8 +64,4 @@ export class ReportF3XService extends ReportService {
       this.store.dispatch(setCashOnHandAction({ payload: payload }));
     }
   }
-
-  public startAmendment(report: Report): Observable<string> {
-    return this.apiService.post(`${this.apiEndpoint}/${report.id}/amend/`, {});
-  }
 }
